move ray tuts
