+++ tutorials/Ray/render_rllib_pistonball.py
@@ -10,9 +10,9 @@ from ray.rllib.agents.ppo import PPOTrainer
 from ray.rllib.env.wrappers.pettingzoo_env import PettingZooEnv
 from ray.rllib.models import ModelCatalog
 from ray.tune.registry import register_env
-from rllib_pistonball import CNNModelV2
 
 from pettingzoo.butterfly import pistonball_v6
+from tutorials.Ray.rllib_pistonball import CNNModelV2
 
 os.environ["SDL_VIDEODRIVER"] = "dummy"
